Fix segments not reappearing on re-selection
After being toggled off, then toggled on, segments in the "sites logged in" and "sign-in
attempts" reports weren't shown.

The issue was occurring because, in the process of updating displayed
segments, the series in the report object were being overwritten with the
"filtered" series — the ones with the segment removed. Thus, this data
was being lost permanently.

To fix this, the full (unfiltered) series are stored in a temporary
variable before the graph is updated; then they are restored.

This regression was introduced in fa809ba.