Report percentage of successful new users
This report is identical to the "sites" report (number of sites logged
in), except for one difference. While the "sites" report lets users
switch between "area" and "line" mode for the graph display, this report
loads into "line" mode by default and hides the option to switch modes.

This was done because area mode is cumulative, which doesn't make sense
with fractions that this report uses.
(e.g., if we had a segmentation with three browsers, each at 0.5 success, the
cumulative graph would display 1.5=150% success, which is nonsensical.)

Closes #15